refactor(Dialog): useBodyScrollLock を DialogContentInner にインライン化
useBodyScrollLock の利用箇所が DialogContentInner のみのため、
別ファイルに切り出す意味がないのでインライン化する
